WIP on parser outline

+++ src/lexer/machines/expr_end.rs
@@ -246,7 +246,7 @@ pub fn construct_machine_expr_end( patterns: &TMatchingPatterns, shared_actions:
             format!(r"[1-9][[:digit:]]*_?{}", pattern_lit!("int_dec")),
             |lexer: &mut Lexer| {
                 // TODO DUMMY
-                let int = lexer.input_stream.current_token().unwrap().parse::<i64>().unwrap();
+                let int = lexer.input_stream.current_token().unwrap().parse::<isize>().unwrap();
                 lexer.emit_token(Token::T_INTEGER(int));
                 lexer.flag_breaking();
             }
